requirements.txt
beautifulsoup4==4.12.3
pandas==2.2.2
python-dotenv==1.0.1
Requests==2.32.3
selenium==4.21.0
supabase==2.5.1
